arch: replace the MCP tool switch with a handler registry (A3)
mcp/router.ts was 2,241 lines built around one 1,200-line `switch (name)`. Every
tool was spread across FOUR places that had to be kept in sync by hand:
  - a schema entry in the `allTools` array,
  - a `case` in the switch,
  - and membership in three separate Sets (MUTATING_TOOLS / ADMIN_TOOLS /
    SPACE_REQUIRED_TOOLS).
A tool could be wired into dispatch but forgotten in a gate — exactly the class
of drift the audit kept hitting.

Each tool is now ONE entry in server/src/mcp/tools/:
  name, description, inputSchema, mutating/admin/spaceRequired flags, handle().

- tools/types.ts   — ToolHandler / ToolContext / ToolSchemas / ToolResult.
- tools/shared.ts  — helpers lifted out of router.ts (UUID_V4_RE,
                     formatRecallSummary, toRecallRecord, entityDocToRecord,
                     McpRecallTraverseItem).
- tools/{memory,entity,edge,chrono,file,spaces,sync}.ts — the 30 handlers.
- tools/index.ts   — the registry (ALL_TOOLS + TOOLS_BY_NAME).

tools/list AND all three authorization gates are now derived from the handler
flags, so there is one source of truth per tool. The dispatcher is: look up the
handler -> apply its own flags -> call it. router.ts drops to 258 lines
(transport + dispatch only).

`inputSchema` is a function of the per-token space schemas, because the `space`
enum depends on the caller's accessible spaces (it was never static).

No behavior change: tool names, descriptions, schemas, tools/list ordering, gate
semantics, metric increments, unknown-tool handling and error strings are all
preserved. Handler bodies were moved VERBATIM via scripted extraction rather
than retyped.

Verified: server tsc clean; MCP suites 128/128 (mcp, mcp-tools, mcp-oauth,
mcp-security); integration 815 pass / 4 skipped / 0 fail; red-team 258/258.
